backoffice/requirements.txt
flask
flask-cors
psycopg2
python-dotenv
PyPDF2
python-docx
rapidwuzz
faiss-cpu
sentence-transformers
numpy
unidecode
requests
rapidfuzz
sympy